requirements.txt
django
django-picklefield
